+++ dwds/lib/src/debugging/metadata/function.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.import 'dart:async';
 
+// @dart = 2.9
+
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
 import '../../loaders/strategy.dart';