+++ dwds/lib/src/readers/asset_reader.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 /// A reader for Dart sources and related source maps.
 abstract class AssetReader {
   /// Returns the contents for a source map at the provided server path, or