+++ dwds/lib/src/utilities/objects.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.import 'dart:async';
 
+// @dart = 2.9
+
 /// A library for WebKit mirror objects and support code. These probably should
 /// get migrated into webkit_inspection_protocol over time.