+++ frontend_server_common/lib/src/package_map.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart = 2.9
+
 // Note: this is a copy from flutter tools
 
 // TODO(annagrin): unify this code so both repos can use it