+++ dwds/test/debug_extension_test.dart
@@ -7,6 +7,7 @@
 // When run locally this test may require a manifest key. This makes it easy to
 // just skip it.
 @Tags(['extension'])
+@Timeout(Duration(seconds: 60))
 @OnPlatform({
   'windows': Skip('https://github.com/dart-lang/webdev/issues/711'),
 })
@@ -51,13 +52,19 @@ void main() async {
         });
 
         test('can close DevTools and relaunch', () async {
-          await (await context.webDriver.windows.toList()).last.close();
+          for (var window in await context.webDriver.windows.toList()) {
+            await context.webDriver.driver.switchTo.window(window);
+            if (await context.webDriver.title == 'Dart DevTools') {
+              await window.close();
+              break;
+            }
+          }
 
           // Relaunch DevTools by (fake) clicking the extension.
           await context.extensionConnection.sendCommand('Runtime.evaluate', {
             'expression': 'fakeClick()',
           });
-          await Future.delayed(const Duration(seconds: 2));
+          await Future.delayed(const Duration(seconds: 4));
           var windows = await context.webDriver.windows.toList();
           await context.webDriver.driver.switchTo.window(windows.last);
           expect(await context.webDriver.title, 'Dart DevTools');