+++ dwds/CHANGELOG.md
@@ -10,6 +10,8 @@
 - Add `DwdsEvent`s around stepping and evaluation.
 - Send an event to the Dart Debug Extension that contains VM service protocol URI. 
 - Depend on `package:vm_service` version `6.1.0+1`.
+- Update the `keepAlive` configs to prevent accidental reuse of a connection after stopping
+  a debug session.
 
 **Breaking changes:**
 - `LoadStrategy`s now require a `moduleInfoForEntrypoint`.